+++ src/main/java/jackiecrazy/footwork/event/MeleeKnockbackEvent.java
@@ -1,20 +1,20 @@
 package jackiecrazy.footwork.event;
 
-import net.minecraft.entity.LivingEntity;
-import net.minecraft.util.DamageSource;
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.damagesource.DamageSource;
 import net.minecraftforge.event.entity.living.LivingEvent;
 import net.minecraftforge.eventbus.api.Event;
 
 @Event.HasResult
 public class MeleeKnockbackEvent extends LivingEvent {
-    protected final float originalStrength;
+    protected final double originalStrength;
     protected final double originalRatioX, originalRatioZ;
     protected LivingEntity attacker;
     protected DamageSource ds;
-    protected float strength;
+    protected double strength;
     protected double ratioX, ratioZ;
 
-    public MeleeKnockbackEvent(LivingEntity attacker, DamageSource source, LivingEntity target, float strength, double ratioX, double ratioZ) {
+    public MeleeKnockbackEvent(LivingEntity attacker, DamageSource source, LivingEntity target, double strength, double ratioX, double ratioZ) {
         super(target);
         this.strength = this.originalStrength = strength;
         this.ratioX = this.originalRatioX = ratioX;
@@ -23,7 +23,7 @@ public class MeleeKnockbackEvent extends LivingEvent {
         this.attacker = attacker;
     }
 
-    public float getStrength() {return this.strength;}
+    public double getStrength() {return this.strength;}
 
     public void setStrength(float strength) {this.strength = strength;}
 
@@ -35,7 +35,7 @@ public class MeleeKnockbackEvent extends LivingEvent {
 
     public void setRatioZ(double ratioZ) {this.ratioZ = ratioZ;}
 
-    public float getOriginalStrength() {return this.originalStrength;}
+    public double getOriginalStrength() {return this.originalStrength;}
 
     public double getOriginalRatioX() {return this.originalRatioX;}
